Temporary fix for #18
This fixes #8 because it no longer relies on `.disabled()` as it uses `Introspect` instead. This is only a temporary solution, as i don't want to rely on other packages and as im planning to port my own UIScrollView to SwiftUI. The custom port would fix this issue and could make other features on my ToDo List (see #11) possible.